Remove dependency and move helper to the right spot.
