+++ airbyte-server/src/main/java/io/airbyte/server/converters/ApiPojoConverters.java
@@ -16,7 +16,7 @@ import io.airbyte.commons.enums.Enums;
 import io.airbyte.config.JobSyncConfig.NamespaceDefinitionType;
 import io.airbyte.config.Schedule;
 import io.airbyte.config.StandardSync;
-import io.airbyte.workers.helper.CatalogConverter;
+import io.airbyte.server.handlers.helpers.CatalogConverter;
 import java.util.stream.Collectors;
 
 public class ApiPojoConverters {
