+++ airbyte-server/src/test/java/io/airbyte/server/converters/CatalogConverterTest.java
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import io.airbyte.commons.enums.Enums;
 import io.airbyte.config.DataType;
 import io.airbyte.server.helpers.ConnectionHelpers;
-import io.airbyte.workers.helper.CatalogConverter;
+import io.airbyte.server.handlers.helpers.CatalogConverter;
 import org.junit.jupiter.api.Test;
 
 class CatalogConverterTest {
