+++ airbyte-workers/build.gradle
@@ -20,7 +20,6 @@ dependencies {
     implementation 'org.eclipse.jetty:jetty-servlet:9.4.31.v20200723'
 
     implementation project(':airbyte-analytics')
-    implementation project(':airbyte-api')
     implementation project(':airbyte-commons-docker')
     implementation project(':airbyte-config:models')
     implementation project(':airbyte-config:persistence')
